[REF] website_sale_product_contract_gift: use new js system
