+++ website_sale_product_contract_gift/__manifest__.py
@@ -24,7 +24,7 @@
     ],
     "assets": {
         "web.assets_frontend": [
-            "website_sale_product_contract_gift/static/src/js/website_sale.js",
+            "website_sale_product_contract_gift/static/src/js/website_sale.esm.js",
         ],
     },
     "demo": [],
